back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "LLM destinated to arm the most basic version for autonomo 1 UIDE."
authors = [
    {name = "Guillermo Andrade",email = "guillermo@quantumpay.mx"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
